Fix bug - platform_tile has wrong coordinates when initialized
Cause: The Tiled software's coordinate system uses the top-left corner convention, but our program uses the center-center one, in fact, all tile and object in the game is affected by this different. But because ALL OF THEM are affected, and also because most of them have same sizes, there doesn't seem to be a problem visually, only the left edge of the map is cropped a bit. The problem occurs when there's an object with different sizes from most of other objects like the platform_tile in this case.

Solution: Adjust the platform_tile coordinates a bit when intializing

TODO: Find some way to fix this whole incompatibility