Create items for brick_questionmark at run-time
This is done by letting CMario class holding a CONST reference to current scene and at run-time add the items to that "objects" vector

TODO: Because the way the Render() method is implemented in CPlayScene, the objects that appear first in "objects" vector will be rendered first. Since the items are added at run-time, they will always at the end of the vector and rendered last which makes them cover the containing brick. Possibly add some ways to sort the vector