Fix bug - fireshot fired by plant_red_fire travels too fast sometimes
Cause: Fireshot's vx is constant and vy is adjusted based on vx to create shooting angle. Then when mario is close to the plant (vx would be too high for this distance) and if the angle is high (mario jumps), vy has to be super high to match vx and create that angle. Then we have a light-speed vy

Solution: Adjust vx based on mario's distance to the plant, then adjust vy based on vx. That way vx will lowers when mario is close to the plant, so vy does not need to be too high to match vx