Fix bug - mario slips through platforms
Cause: The platforms were not considered within camera range. Basically, for the platforms to be in range, they need to have their grid info updated. But for the platforms to have their grid info updated, they need to be in range.

Solution: Add all objects into coObjects vector, so that even if the platforms are not considered within camera range, mario can still collide with them and stand on them